+++ CHANGELOG.markdown
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 ### Added
 
 * Added support for the new `op sign` and `unpackcolor` Mindustry BE instructions to the compiler and processor emulator.
+* Added support for using `null` literals in the `case` expression.
 
 ### Changed